Added AI preception systrem and detection
Usage
Add to Behavior Tree:
Add an instance of UBTTask_FindPlayerLocation to your behavior tree.
Configuration:
Configure the task by setting the appropriate Blackboard key and adjusting parameters such as SearchRandom and SearchRadius.
Execution:
The task will find the player's location and store it in the Blackboard during execution.
Parameters
SearchRandom: If true, the task will find a random navigable location near the player. If false, it will use the player's exact location.
SearchRadius: The radius within which to search for a random navigable location.
Dependencies
UNavigationSystemV1
UGameplayStatics